Move CacheVC struct and methods from Vol
